0_WSI_alignment_and_normalisation/requirements_for_visualisation_env.txt
absl-py==1.0.0
affine==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anndata==0.9.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Automat==0.8.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==6.1.0
blinker==1.4
boto3==1.35.73
botocore==1.35.73
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.4
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloud-init==23.1.2
cloudpickle==3.0.0
colorama==0.4.3
coloredlogs==7.3
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.1.1
cryptography==2.8
ctk-cli==1.5
cupshelpers==1.0
cutadapt==2.8
cwltool==2.0.20200224214940
cycler==0.12.1
dask==2023.5.0
dask-image==2023.3.0
dbus-python==1.2.16
debugpy==1.8.5
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dicomweb-client==0.59.3
diskcache==5.6.3
distlib==0.3.8
distributed==2023.5.0
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.8
dnaio==0.4.1
docrep==0.3.2
entrypoints==0.3
et-xmlfile==1.1.0
etils==1.3.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
flax==0.7.2
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gast==0.4.0
GDAL==3.3.2
get-annotations==0.1.2
girder-client==3.2.6
girder-slicer-cli-web==1.5.11
Glymur==0.13.6
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
h5py==3.6.0
histomicstk==1.3.14
html5lib==1.0.1
httpcore==1.0.5
httplib2==0.14.0
httpx==0.27.2
humanfriendly==4.18
hyperlink==19.0.0
idna==2.8
igraph==0.11.6
imagecodecs==2023.3.16
imagecodecs-numcodecs==2023.3.16
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
inflect==7.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isodate==0.6.0
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keras==2.13.1
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.7
kornia-rs==0.1.8
language-selector==0.1
large-image==1.30.3
large-image-converter==1.30.3
large-image-source-bioformats==1.30.3
large-image-source-deepzoom==1.30.3
large-image-source-dicom==1.30.3
large-image-source-dummy==1.30.3
large-image-source-gdal==1.30.3
large-image-source-mapnik==1.30.3
large-image-source-multi==1.30.3
large-image-source-nd2==1.30.3
large-image-source-ometiff==1.30.3
large-image-source-openjpeg==1.30.3
large-image-source-openslide==1.30.3
large-image-source-pil==1.30.3
large-image-source-rasterio==1.30.3
large-image-source-test==1.30.3
large-image-source-tiff==1.30.3
large-image-source-tifffile==1.30.3
large-image-source-vips==1.30.3
large-image-source-zarr==1.30.3
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.4
leidenalg==0.10.2
libclang==18.1.1
lightning-utilities==0.11.7
llvmlite==0.38.1
locket==1.0.0
lockfile==0.12.2
lxml==4.5.0
macaroonbakery==1.3.1
mapnik==4.0.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mdurl==0.1.2
mistune==3.0.2
ml-collections==0.1.1
ml-dtypes==0.2.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mudata==0.2.3
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==0.4.3
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nd2==0.10.1
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nimfa==1.4.0
notebook==7.2.2
notebook-shim==0.2.4
numba==0.55.2
numcodecs==0.12.1
numpy==1.22.1
numpyro==0.12.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
ome-types==0.5.3
omnipath==1.0.8
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
openslide-bin==4.0.0.6
openslide-python==1.4.1.1
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint==0.2.3
overrides==7.7.0
packaging==24.1
palettable==3.3.3
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
patsy==0.5.6
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
PIMS==0.7
pkgconfig==1.5.5
pkgutil-resolve-name==1.3.10
platformdirs==4.3.2
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==4.25.5
prov==1.5.2
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pydantic==2.10.2
pydantic-compat==0.1.2
pydantic-core==2.27.1
pydicom==2.4.4
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylibtiff==0.6.1.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynndescent==0.5.13
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==3.5.0
pyRFC3339==1.1
pyro-api==0.1.2
pyro-ppl==1.9.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-bioformats==4.1.0.8.0.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-igraph==0.11.6
python-javabridge==4.0.4
python-json-logger==2.0.7
pytorch-lightning==1.9.5
pytz==2024.2
pyvips==2.2.3
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.3.10
rdflib==4.2.2
rdflib-jsonld==0.4.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
resource-backed-dask-array==0.1.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.1
rpds-py==0.20.0
rsa==4.8
ruamel.yaml==0.15.89
s3transfer==0.10.4
scanpy==1.9.8
schema-salad==5.0.20200220195218
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
scvi-tools==0.20.3
seaborn==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.3
service-identity==18.1.0
session-info==1.0.0
shapely==2.0.6
shellescape==3.4.1
simplejson==3.16.0
six==1.14.0
slicerator==1.1.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
sos==4.4
soupsieve==1.9.5
SPARQLWrapper==1.8.5
SQLAlchemy==2.0.36
squidpy==1.2.3
ssh-import-id==5.10
stack-data==0.6.3
staintools==2.1.2
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.13.3
systemd-python==234
tblib==3.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1
tensorrt-dispatch==8.6.1
tensorrt-lean==8.6.1
tensorstore==0.1.45
termcolor==1.1.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tifftools==1.5.2
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
torch==2.4.1
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
Twisted==18.9.0
typeguard==4.3.0
types-python-dateutil==2.9.0.20240906
typing-extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
umap-learn==0.5.6
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.20
validators==0.34.0
virtualenv==20.26.5
wadllib==1.3.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.2
widgetsnbextension==4.0.13
wrapt==1.13.3
wsidicom==0.13.0
xarray==2023.1.0
xkit==0.0.0
xopen==0.8.4
xsdata==24.3.1
yarl==1.11.1
zarr==2.16.1
zict==3.0.0
zipp==3.20.1
zope.interface==4.7.1
